+++ cpg-concepts/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/concepts/auth/Authentication.kt
@@ -23,7 +23,7 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.graph.concepts.iam
+package de.fraunhofer.aisec.cpg.graph.concepts.auth
 
 import de.fraunhofer.aisec.cpg.graph.Node
 import de.fraunhofer.aisec.cpg.graph.concepts.Concept
@@ -31,8 +31,8 @@ import de.fraunhofer.aisec.cpg.graph.concepts.Operation
 import de.fraunhofer.aisec.cpg.graph.edges.flows.EvaluationOrder
 import java.util.Objects
 
-/** Represents a high-level concept for identity and access management (IAM). */
-abstract class IdentityAccessManagement(underlyingNode: Node?) : Concept(underlyingNode)
+/** Represents a high-level concept for authentication. */
+abstract class Authentication(underlyingNode: Node?) : Concept(underlyingNode)
 
 /**
  * Represents a token-based authentication.
@@ -40,7 +40,7 @@ abstract class IdentityAccessManagement(underlyingNode: Node?) : Concept(underly
  * @param token The authentication token, which may be an opaque token.
  */
 open class TokenBasedAuth(underlyingNode: Node? = null, val token: Node) :
-    IdentityAccessManagement(underlyingNode) {
+    Authentication(underlyingNode) {
     override fun equals(other: Any?): Boolean {
         return other is TokenBasedAuth && super.equals(other) && other.token == this.token
     }
@@ -67,7 +67,7 @@ class JwtAuth(underlyingNode: Node? = null, val jwt: Node, val payload: Node) :
 }
 
 /** Abstract base class for authentication operations. */
-abstract class IAMOperation(underlyingNode: Node? = null, concept: IdentityAccessManagement) :
+abstract class AuthenticationOperation(underlyingNode: Node? = null, concept: Authentication) :
     Operation(underlyingNode, concept)
 
 /**
@@ -76,11 +76,8 @@ abstract class IAMOperation(underlyingNode: Node? = null, concept: IdentityAcces
  * @param credential The credential can be a call (e.g., a function call that reads a header) or a
  *   variable that holds the value, e.g. the token
  */
-class Authenticate(
-    underlyingNode: Node? = null,
-    concept: IdentityAccessManagement,
-    val credential: Node,
-) : IAMOperation(underlyingNode, concept) {
+class Authenticate(underlyingNode: Node? = null, concept: Authentication, val credential: Node) :
+    AuthenticationOperation(underlyingNode, concept) {
     override fun equals(other: Any?): Boolean {
         return other is Authenticate && super.equals(other) && other.credential == this.credential
     }
@@ -93,7 +90,7 @@ class Authenticate(
  *
  * @param jwt The JWT containing encoded authentication information.
  */
-class IssueJwt(underlyingNode: Node, jwt: JwtAuth) : IAMOperation(underlyingNode, jwt) {
+class IssueJwt(underlyingNode: Node, jwt: JwtAuth) : AuthenticationOperation(underlyingNode, jwt) {
     override fun equals(other: Any?): Boolean {
         return other is IssueJwt && super.equals(other)
     }
@@ -106,7 +103,8 @@ class IssueJwt(underlyingNode: Node, jwt: JwtAuth) : IAMOperation(underlyingNode
  *
  * @param jwt The JWT containing encoded authentication information.
  */
-class ValidateJwt(underlyingNode: Node, jwt: JwtAuth) : IAMOperation(underlyingNode, jwt) {
+class ValidateJwt(underlyingNode: Node, jwt: JwtAuth) :
+    AuthenticationOperation(underlyingNode, jwt) {
     override fun equals(other: Any?): Boolean {
         return other is ValidateJwt && super.equals(other)
     }
@@ -119,7 +117,8 @@ class ValidateJwt(underlyingNode: Node, jwt: JwtAuth) : IAMOperation(underlyingN
  *
  * @param jwt The JWT containing encoded authentication information.
  */
-class AuthorizeJwt(underlyingNode: Node, jwt: JwtAuth) : IAMOperation(underlyingNode, jwt) {
+class AuthorizeJwt(underlyingNode: Node, jwt: JwtAuth) :
+    AuthenticationOperation(underlyingNode, jwt) {
     /** The next EOG edges which are followed if the authorization was successful. */
     val nextGrantedEOGEdges: List<EvaluationOrder>
         get() = nextEOGEdges.filter { it.branch == true }
