+++ cpg-concepts/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/concepts/http/HttpClient.kt
@@ -28,14 +28,14 @@ package de.fraunhofer.aisec.cpg.graph.concepts.http
 import de.fraunhofer.aisec.cpg.graph.Node
 import de.fraunhofer.aisec.cpg.graph.concepts.Concept
 import de.fraunhofer.aisec.cpg.graph.concepts.Operation
-import de.fraunhofer.aisec.cpg.graph.concepts.iam.IdentityAccessManagement
+import de.fraunhofer.aisec.cpg.graph.concepts.auth.Authentication
 import java.util.Objects
 
 /** Represents an [HttpClient]. */
 class HttpClient(
     underlyingNode: Node? = null,
     val isTLS: Boolean? = false,
-    val authentication: IdentityAccessManagement? = null,
+    val authentication: Authentication? = null,
 ) : Concept(underlyingNode = underlyingNode) {
     override fun equals(other: Any?): Boolean {
         return other is HttpClient &&
